+++ zititest/ziti-fabric-test/subcmd/loop3/protocol.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/foundation/v2/info"
-	"github.com/openziti/ziti/ziti-fabric-test/subcmd/loop3/pb"
+	"github.com/openziti/ziti/zititest/ziti-fabric-test/subcmd/loop3/pb"
 	"github.com/pkg/errors"
 	"google.golang.org/protobuf/proto"
 	"io"